fix: update inubekit library
