+++ package.json
@@ -16,7 +16,7 @@
   },
   "dependencies": {
     "@auth0/auth0-react": "^2.2.4",
-    "@inubekit/inubekit": "^3.0.0",
+    "@inubekit/inubekit": "^4.0.0",
     "@isettingkit/business-rules": "^1.0.2",
     "@isettingkit/input": "^1.0.1",
     "@isettingkit/view": "^0.1.23",
