+++ src/design/feedback/RequestProcess/stories/RequestProcess.stories.tsx
@@ -2,9 +2,9 @@ import { BrowserRouter } from "react-router-dom";
 import { Meta, StoryFn } from "@storybook/react";
 import { useState } from "react";
 import { Button } from "@inubekit/inubekit";
-
 import { requestStatusMessage } from "@config/moneyDestination/moneyDestinationTab/generics/requestStatusMessage";
-import { IRequestProcess, RequestProcess } from "..";
+import { IRequestProcessContent } from "@ptypes/design/IRequestProcessContent";
+import { RequestProcess } from "..";
 
 const meta: Meta<typeof RequestProcess> = {
   title: "feedback/RequestProcess",
@@ -30,7 +30,7 @@ const InterventionHumanData = {
   requestStatus: "PendingApproval",
 };
 
-const Template: StoryFn<IRequestProcess> = (args) => {
+const Template: StoryFn<IRequestProcessContent> = (args) => {
   const [showModal, setShowModal] = useState(false);
 
   return (
