--- src/design/modals/requestProcessModal/types.ts
@@ -1,7 +1,0 @@
-interface IRequestSteps {
-  name: string;
-  status: "pending" | "completed" | "error";
-  statusRequest?: string;
-}
-
-export type { IRequestSteps };
