+++ src/hooks/moneyDestination/useSaveMoneyDestination/index.ts
@@ -2,7 +2,6 @@ import { useNavigate } from "react-router-dom";
 import { useState, useEffect, useContext } from "react";
 import { IFlagAppearance, useFlag } from "@inubekit/inubekit";
 import { statusFlowAutomatic } from "@config/status/statusFlowAutomatic";
-import { IRequestSteps } from "@design/modals/requestProcessModal/types";
 import { ISaveDataResponse } from "@ptypes/saveData/ISaveDataResponse";
 import { flowAutomaticMessages } from "@config/moneyDestination/moneyDestinationTab/generics/flowAutomaticMessages";
 import { interventionHumanMessage } from "@config/moneyDestination/moneyDestinationTab/generics/interventionHumanMessage";
@@ -19,6 +18,7 @@ import { RequestStepsStatus } from "@enum/requestStepsStatus";
 import { postSaveRequest } from "@services/requestInProgress/postSaveRequest";
 import { ChangeToRequestTab } from "@context/changeToRequestTab/changeToRequest";
 import { requestStepsInitial } from "@config/requestSteps";
+import { IRequestSteps } from "@ptypes/design/IRequestSteps";
 
 const useSaveMoneyDestination = (props: IUseSaveMoneyDestination) => {
   const {
