+++ src/types/design/IRequestProcess/index.ts
@@ -1,0 +1,10 @@
+import { IRequestSteps } from "../IRequestSteps";
+
+interface IRequestProcess {
+  requestSteps: IRequestSteps[];
+  sizeIcon: string;
+  stepCurrent: number;
+  stepCurrentIndex: number;
+}
+
+export type { IRequestProcess };
